pyproject.toml
[project]
name = "nonebot-plugin-batitle"
version = "0.1.4"
description = "BlueArchive Title Generator"
authors = [
    {name = "MerCuJerry", email = "mercujerry@gmail.com"},
]
dependencies = [
    "nonebot2 >= 2.1.0",
    "nonebot-adapter-onebot>=2.3.1",
    "pillow <10.0.0, >=9.2.0"
]
requires-python = ">=3.8,<4.0"
readme = "README.md"
license = {file = "LICENSE"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.urls]
homepage = "https://github.com/MerCuJerry/nonebot-plugin-batitle"

[tool.pdm.build]
includes = ["nonebot_plugin_batitle"]
